pyproject.toml
[project]
name = "pyfltr"
dynamic = ["version"]
description = "Python Formatters, Linters, and Testers Runner."
authors = [{ name = "aki.", email = "aki@tqzh.tk" }]
requires-python = ">=3.11,<4.0"
readme = "README.md"
license = {text = "MIT"}
classifiers = [
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "mcp>=1.0",
    "natsort>=8.4.0",
    "platformdirs>=4.0",
    "pre-commit>=4.5.1",
    "psutil>=6.0",
    "python-ulid>=3.0",
    "pyyaml>=6.0.3",
    "textual>=8.0.0",
    "tomlkit>=0.13",
    "typos>=1.0",
    "yamllint>=1.0",
]

[project.optional-dependencies]
# Python 系 linter / tester 依存を導入するための extras。
# v3.0.0 でオプトイン化したため、Python プロジェクトでは
# ``pip install pyfltr[python]`` または ``uv add 'pyfltr[python]'`` で追加する。
python = [
    "dill>=0.3",
    "mypy>=1.0",
    "pylint-pydantic>=0.4.1",
    "pylint>=3.0",
    "pyright[nodejs]>=1.1.403",
    "pytest-asyncio>=1.0.0",
    "pytest>=7.0",
    "ruff>=0.12.1",
    "ty>=0.0.26",
    "uv-sort>=0.5",
]

[project.urls]
Homepage = "https://github.com/ak110/pyfltr"

[project.scripts]
pyfltr = "pyfltr.cli.main:main"

[dependency-groups]
# pyfltr 自体が Python 製のため、開発時は Python 系ツールを自身の extras 経由で導入する。
# これにより本体リリース時に Python 系ツール依存を抱え込まない構造を保ちつつ、
# 開発・CI では ``uv sync --group dev`` 一発で linter / tester を揃えられる。
dev = [
    "mkdocs-llmstxt>=0.2",
    "mkdocs-material>=9.6",
    "mkdocs>=1.6",
    "pyfltr[python]",
    "pytest-mock",
    "types-pyyaml>=6.0.12.20260408",
]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.uv]
exclude-newer = "1 day"

[tool.pyfltr]
preset = "latest"
# preset = "latest" + python = true で Python 系推奨構成 (ruff-format / ruff-check /
# mypy / pylint / pyright / pytest / uv-sort) が gate 通過で有効化される。
# ``ty`` のみ preset 非収録のため個別 true で追加する。
python = true
ty = true
glab-ci-lint = true
yamllint = true
taplo = true
gitleaks = true
ec = true
shellcheck = true
shfmt = true
hadolint = true
pylint-args = ["--jobs=4"]

[tool.ruff]
# https://docs.astral.sh/ruff/configuration/
line-length = 128

[tool.ruff.lint]
# https://docs.astral.sh/ruff/linter/#rule-selection
select = [
    # pydocstyle
    "D",
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # flake8-import-conventions
    "ICN",
    # isort
    "I",
]
ignore = [
    "D107", # Missing docstring in `__init__`
    "D415", # First line should end with a period
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"**_test.py" = ["D"]
"**/__init__.py" = ["D104"]  # Missing docstring in public package

[tool.mypy]
# https://mypy.readthedocs.io/en/stable/config_file.html
allow_redefinition = true
check_untyped_defs = true
ignore_missing_imports = true
strict_optional = true
strict_equality = true
warn_no_return = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true

[tool.ty.rules]
# Windows 限定の attr-defined 誤検知を抑止する `# ty: ignore` コメントが、
# Linux 上では「未使用」扱いとなり警告になってしまうため、プラットフォーム跨ぎの
# POSIX 分岐を扱うファイルで恒常的に衝突する。ty: ignore を一元無効化する。
unused-ignore-comment = "ignore"

[tool.typos.default.extend-words]
# テストデータ内の意図的なtypo
teh = "teh"
hte = "hte"

[tool.pytest.ini_options]
# https://docs.pytest.org/en/latest/reference/reference.html#ini-options-ref
addopts = "--showlocals -p no:cacheprovider --maxfail=5 --durations=30 --durations-min=0.5"
log_level = "DEBUG"
xfail_strict = true
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
